+++ server/Controllers/ForumController.js
@@ -1,5 +1,44 @@
 import Forum from "../Models/Forum.js";
 import Category from "../Models/Category.js";
+import Thread from "../Models/Thread.js";
+
+export const followForum = async (req, res) => {
+
+    const forumId = req.params.id;
+    const type = req.params.type;
+    const userId = req.body.userId;
+
+    try {
+        const forum = await Forum.findById(forumId);
+        console.log(forum)
+        if (type === "0" && forum.followers.includes(userId)) {
+            // unfollow
+            await forum.updateOne({ $pull: { followers: userId } });
+        } else if (type === "1" && !forum.followers.includes(userId)) {
+            // follow
+            await forum.updateOne({ $push: { followers: userId } });
+        } else {
+            return res.status(403).json({ message: "Forbidden action." })
+        }
+        
+        return res.status(200).json({ message: "Follow status updated!" });
+    } catch (err) {
+        return res.status(500).json({ message: err });
+    }
+};
+
+export const getForumById = async (req, res) => {
+    const forumId = req.params.id;
+
+    try {
+        const forum = await Forum.findById(forumId);
+
+        return res.status(200).json(forum);
+    } catch (err) {
+        console.log(err);
+        return res.status(500).json({ message: err });
+    }
+}
 
 export const createForum = async (req, res) => {
 
@@ -42,10 +81,16 @@ export const deleteForum = async (req, res) => {
 
     try {
         
-        await Category.findByIdAndUpdate({ _id: categoryId }, { $pull: { forums: forumId } });
-        await Forum.findOneAndDelete({ _id: forumId });
-        
-        return res.status(200).json({ message: "Forum deleted." });
+        const threads = await Thread.find({ forumId: forumId });
+
+        if (!threads) {
+            await Category.findByIdAndUpdate({ _id: categoryId }, { $pull: { forums: forumId } });
+            await Forum.findOneAndDelete({ _id: forumId });
+            return res.status(200).json({ message: "Forum deleted." });
+
+        }
+        return res.status(403).json({ message: "Action forbidden. Threads of " + threads.length + " remaining." });
+
     } catch (err) {
         console.log(err);
         return res.status(500).json({ message: err });
